fix a big bug in read_head function
This function find a pair ']' brace for current '[' brace in code array.
But it was that finction try find pair with check a cell of input with
position = *code_pos in cycle, not a code array.
It is not right. Need check code array with index of cell = *code_pos